+++ src/main/java/kh/bookday/dao/WishlistDAO.java
@@ -1,6 +1,7 @@
 package kh.bookday.dao;
 
 import java.util.List;
+import java.util.Map;
 
 import org.apache.ibatis.session.SqlSession;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -16,5 +17,15 @@ public class WishlistDAO {
 	
 	public List<WishlistDTO> selectWishlistListById(String id){
 		return db.selectList("Wishlist.selectWishlistListById", id);
+	}
+	
+	/* 위시리스트 체크 */
+	public WishlistDTO selectWishlistByIdBisbn(Map<String, String> param) {
+		return db.selectOne("Bookbag.selectWishlistByIdBisbn", param);
+	}
+	
+	/* 위시리스트 추가 */
+	public void insertWishlist(WishlistDTO dto) {
+		db.insert("Bookbag.insertWishlist", dto);
 	}
 }
