+++ src/main/java/kh/bookday/service/WishlistService.java
@@ -1,6 +1,8 @@
 package kh.bookday.service;
 
+import java.util.HashMap;
 import java.util.List;
+import java.util.Map;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -16,5 +18,18 @@ public class WishlistService {
 	
 	public List<WishlistDTO> selectWishlistListById(String id){
 		return dao.selectWishlistListById(id);
+	}
+	
+	/* 위시리스트 체크 */
+	public WishlistDTO selectWishlistByIdBisbn(String id, String b_isbn) {
+		Map<String, String> param = new HashMap<>();
+		param.put("id", id);
+		param.put("b_isbn", b_isbn);
+		return dao.selectWishlistByIdBisbn(param);
+	}
+	
+	/* 위시리스트 추가 */
+	public void insertWishlist(WishlistDTO dto) {
+		dao.insertWishlist(dto);
 	}
 }
