+++ src/main/webapp/WEB-INF/spring/root-context.xml
@@ -6,8 +6,8 @@
 	<!-- Root Context: defines shared resources visible to all other web components -->
 	<bean id="dataSource" class="org.apache.commons.dbcp2.BasicDataSource">
 		<property name="driverClassName" value="oracle.jdbc.driver.OracleDriver" />
-		<property name="username" value="JINbookday" />
-		<property name="password" value="JINbookday" />
+		<property name="username" value="bookday" />
+		<property name="password" value="bookday" />
 		<property name="url" value="jdbc:oracle:thin:@localhost:1521:xe" />
 	</bean>
